Remove CameraPage Initialization composition templates and add logic to page templates
