+++ templates/Uwp/Pages/Camera.CaliburnMicro/Views/CameraViewPage.xaml.cs
@@ -8,5 +8,17 @@ namespace Param_ItemNamespace.Views
         {
             InitializeComponent();
         }
+
+        protected override async void OnNavigatedTo(NavigationEventArgs e)
+        {
+            base.OnNavigatedTo(e);
+            await cameraControl.InitializeCameraAsync();
+        }
+
+        protected override async void OnNavigatedFrom(NavigationEventArgs e)
+        {
+            base.OnNavigatedFrom(e);
+            await cameraControl.CleanupCameraAsync();
+        }
     }
 }
