--- templates/Uwp/_composition/CaliburnMicro/Page.Camera.InitializeCamera_TabbedPivot/Views/wts.ItemNamePage_searchreplace.xaml.cs
@@ -1,3 +1,0 @@
-﻿public sealed partial class wts.ItemNamePage : Page
-^^^-searchabove-replacebelow-vvv
-public sealed partial class wts.ItemNamePage : Page, IPivotPage
