--- templates/Uwp/_composition/CodeBehind/Page.Camera.InitializeCamera_Blank_SplitView/.template.config/template.json
@@ -1,28 +1,0 @@
-﻿{
-  "author": "Microsoft",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CodeBehind.Page.Camera.InitializeCamera.BlankSplitView",
-  "tags": {
-    "language": "C#",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform": "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$projectType == Blank|SplitView & identity == wts.Page.Camera.CodeBehind"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
