--- templates/Uwp/_composition/CodeBehind/Page.Camera.InitializeCamera_Blank_SplitView/Views/wts.ItemNamePage_postaction.xaml.cs
@@ -1,28 +1,0 @@
-﻿//{[{
-using Windows.UI.Xaml.Navigation;
-using Param_ItemNamespace.Helpers;
-//}]}
-
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class wts.ItemNamePage : Page, INotifyPropertyChanged
-    {
-        public wts.ItemNamePage()
-        {
-        }
-
-        //{[{
-        protected override async void OnNavigatedTo(NavigationEventArgs e)
-        {
-            base.OnNavigatedTo(e);
-            await cameraControl.InitializeCameraAsync();
-        }
-
-        protected override async void OnNavigatedFrom(NavigationEventArgs e)
-        {
-            base.OnNavigatedFrom(e);
-            await cameraControl.CleanupCameraAsync();
-        }
-        //}]}
-    }
-}
