--- templates/Uwp/_composition/CodeBehind/Page.Camera.InitializeCamera_TabbedPivotVB/.template.config/template.json
@@ -1,28 +1,0 @@
-﻿{
-  "author": "Microsoft",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "CodeBehind.Page.Camera.InitializeCamera.TabbedPivot.VB",
-  "tags": {
-    "language": "VisualBasic",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform": "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$projectType == TabbedPivot & identity == wts.Page.Camera.CodeBehind.VB"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
