--- templates/Uwp/_composition/_shared/Page.Camera.InitializeCamera_Blank_SplitViewVB/.template.config/template.json
@@ -1,28 +1,0 @@
-﻿{
-  "author": "Microsoft",
-  "classifications": [
-    "Universal"
-  ],
-  "name": "Page.Camera.InitializeCamera_Blank_SplitView.VB",
-  "tags": {
-    "language": "VisualBasic",
-    "type": "item",
-    "wts.type": "composition",
-    "wts.platform": "Uwp",
-    "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == MVVMBasic|MVVMLight & $projectType == Blank|SplitView & groupidentity == wts.Page.Camera.VB"
-  },
-  "sourceName": "wts.ItemName",
-  "preferNameDirectory": true,
-  "PrimaryOutputs": [],
-  "symbols": {
-    "wts.rootNamespace": {
-      "type": "parameter",
-      "replaces": "Param_RootNamespace"
-    },
-    "wts.itemNamespace": {
-      "type": "parameter",
-      "replaces": "Param_ItemNamespace"
-    }
-  }
-}
