--- templates/Uwp/_composition/_shared/Page.Camera.InitializeCamera_TabbedPivot/Views/wts.ItemNamePage_postaction.xaml.cs
@@ -1,26 +1,0 @@
-﻿//{[{
-using Param_ItemNamespace.Helpers;
-using System.Threading.Tasks;
-//}]}
-
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class wts.ItemNamePage : Page
-    {
-        public wts.ItemNamePage()
-        {
-        }
-
-        //{[{
-        public async Task OnPivotSelectedAsync()
-        {
-            await cameraControl.InitializeCameraAsync();
-        }
-
-        public async Task OnPivotUnselectedAsync()
-        {
-            await cameraControl.CleanupCameraAsync();
-        }
-        //}]}
-    }
-}
